requirements.txt
Flask==3.0.0
openai-whisper==20231117
torch>=2.2.0
ffmpeg-python==0.2.0
Werkzeug==3.0.1

